reduced code, better html structuring, css fixes
