+++ package.json
@@ -8,6 +8,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "lottie-react": "^2.2.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
