+++ src/components/ButtonReset.jsx
@@ -1,34 +1,25 @@
 import styled from 'styled-components'
 
-export default function ButtonReset({ handleReset }) {
-  return (
-    <>
-      <ButtonResetStyled onClick={() => handleReset()}>RESET</ButtonResetStyled>
-    </>
-  )
+export default function ButtonReset({ handleClick, buttonName }) {
+  return <ButtonResetStyled onClick={() => handleClick()}>{buttonName}</ButtonResetStyled>
 }
 
 const ButtonResetStyled = styled.button`
   background-color: rgb(255, 255, 255, 0);
   border: 1px solid #222222;
   border-radius: 8px;
-  box-sizing: border-box;
-  color: #222222;
   cursor: pointer;
   display: inline-block;
-  font-family: Circular, -apple-system, BlinkMacSystemFont, Roboto, 'Helvetica Neue', sans-serif;
   font-size: clamp(2rem, 3vw, 3rem);
   font-weight: 600;
   line-height: 20px;
   outline: none;
   padding: clamp(1rem, 2vw, 1.5rem);
-  position: relative;
   text-align: center;
   text-decoration: none;
   touch-action: manipulation;
   transition: box-shadow 0.2s, -ms-transform 0.1s, -webkit-transform 0.1s, transform 0.1s;
   user-select: none;
-  width: auto;
 
   :focus-visible {
     box-shadow: #222222 0 0 0 2px, rgba(255, 255, 255, 0.2) 0 0 0 4px;
