+++ src/page/Welcome.jsx
@@ -1,4 +1,3 @@
-import Header from '../components/Header'
 import Introduction from '../components/Introduction'
 import LottieAnimationTwo from '../components/LottieAnimationTwo'
 import { Link } from 'react-router-dom'
@@ -7,7 +6,6 @@ import styled from 'styled-components'
 export default function Welcome() {
   return (
     <WelcomeSection to='/bingo'>
-      <Header />
       <LottieAnimationTwo />
       <Introduction />
     </WelcomeSection>
